fix overzealous replace in wrappers pathnames
Wrapper filenames were generated by replacing ".c" with "_private.h".
Unfortunately this matched anywhere in the string, causing problems if
the path contained a directory containing the substring ".c". For
instance, if box64 is compiled in a '.cache' directory (as AUR helpers
like to do, see #144), the directory name was inadvertently mangled
into '_private.hache'.

Fix this by restricting the match to the end of the string.