Enable test for RV64 and ARM64 in CI
If we can pass the environment variable `QEMU_LD_PREFIX=/usr/{riscv64,aarch64}-linux-gnu/lib/`
to ctest, it should also work and be cleaner, but I don't know how to do it as
I am not familiar with cmake.

Note that TEGRAX1 is not currently working because of a mismatch in GLIBC version.
Therefore, we skipped this platform.

Co-authored-by: xctan <xctan@cirno.icu>