add /usr/x86_64-linux-gnu/lib to the shared library search path
On Debian there are the packages libstdc++6-amd64-cross and
libgcc-s1-amd64-cross which provide amd64 shared libraries which are
usually used for cross compiling without multiarch but using a sysroot
prefix instead. Since libgcc_s.so.1 and libstdc++.so.6 shipped by these
packages are fully functional, let box64 use them as a fallback if it
cannot find them in the other paths that it searched before.